Backend/requirements.txt
flask
flask-cors
langchain-community
langchain-google-genai
youtube-transcript-api
faiss-cpu
python-dotenv
langchain